+++ security/main_test.go
@@ -20,11 +20,9 @@ package security_test
 import (
 	"github.com/cockroachdb/cockroach/security"
 	"github.com/cockroachdb/cockroach/security/securitytest"
-	"github.com/cockroachdb/cockroach/util/log"
 )
 
 func init() {
-	log.SetupTestingFlags()
 	ResetTest()
 }
