+++ server/main_test.go
@@ -20,10 +20,8 @@ package server_test
 import (
 	"github.com/cockroachdb/cockroach/security"
 	"github.com/cockroachdb/cockroach/security/securitytest"
-	"github.com/cockroachdb/cockroach/util/log"
 )
 
 func init() {
-	log.SetupTestingFlags()
 	security.SetReadFileFn(securitytest.Asset)
 }
